+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - Chg #390: Remove `yiisoft/json` dependency (@Tigrov)
 - Enh #393: Refactor according changes in `db` package (@Tigrov)
 - New #391: Add `caseSensitive` option to like condition (@vjik)
+- Enh #396: Remove `getCacheKey()` and `getCacheTag()` methods from `Schema` class (@Tigrov)
 
 ## 1.3.0 March 21, 2024